requirements.txt
# Core LLM & Embeddings
openai==1.30.1

# Vector Store
faiss-cpu==1.8.0
numpy==1.26.4

# Document Loading & Processing
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
pypdf==4.2.0
tiktoken==0.7.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Utilities
colorama==0.4.6
rich==13.7.1

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
